third_party/blink/web_tests/external/wpt/permissions-policy/experimental-features/unload-disallowed-subframe.tentative.window-expected.txt

This is a testharness.js-based test.
[FAIL] 'unload' Policy : allowed in main frame but disallowed in subframe
  assert_equals: main: unload in main should be allowed expected true but got false
Harness: the test ran to completion.
